// user-service/generated-sdk/api.ts
export interface ErrorResponse {
    /**
     * 
     * @type {number}
     * @memberof ErrorResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'message'?: string;
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'details'?: string;
}
// ...
export interface LoginUser200Response {
    /**
     * 
     * @type {string}
     * @memberof LoginUser200Response
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof LoginUser200Response
     */
    'userId'?: number;
    /**
     * 
     * @type {string}
     * @memberof LoginUser200Response
     */
    'token'?: string;
}
/**
 * 
 * @export
 * @interface LoginUserRequest
 */
export interface LoginUserRequest {
    /**
     * 
     * @type {string}
     * @memberof LoginUserRequest
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof LoginUserRequest
     */
    'password': string;
}
/**
 * 
 * @export
 * @interface LogoutUser200Response
 */
export interface LogoutUser200Response {
    /**
     * 
     * @type {string}
     * @memberof LogoutUser200Response
     */
    'message'?: string;
}
/**
 * 
 * @export
 * @interface LogoutUserRequest
 */
export interface LogoutUserRequest {
    /**
     * 
     * @type {string}
     * @memberof LogoutUserRequest
     */
    'token': string;
}
/**
 * 
 * @export
 * @interface RegisterUser201Response
 */
export interface RegisterUser201Response {
    /**
     * 
     * @type {string}
     * @memberof RegisterUser201Response
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof RegisterUser201Response
     */
    'userId'?: number;
}